+++ assets/javascripts/kitten/components/cards/crowdfunding-card/styles.js
@@ -16,6 +16,11 @@ const mobileStyles = css`
   overflow: hidden;
   display: block;
 
+  &[href]:focus {
+    outline: ${COLORS.primary4} solid ${pxToRem(2)};
+    outline-offset: ${pxToRem(2)};
+  }
+
   .k-CrowdfundingCard__paddedContainer {
     padding: 0 ${COMPONENT_GUTTER};
   }
@@ -194,24 +199,6 @@ const mobileStyles = css`
     animation-name: ${loadingKeyframes};
   }
 
-  /* Widget-specific styles */
-
-  &.k-CrowdfundingCardWidget {
-    .k-CrowdfundingCard__titleAndDesc {
-      margin-top: ${pxToRem(5)};
-    }
-
-    .k-CrowdfundingCard__titleWrapper {
-      display: flex;
-    }
-
-    .k-CrowdfundingCard__title__dayCounter {
-      line-height: 1;
-      margin-left: calc(2 * ${COMPONENT_GUTTER});
-      margin-top: ${pxToRem(10)};
-    }
-  }
-
   /* isLoading */
 
   &.k-CrowdfundingCard--isLoading {
@@ -341,12 +328,6 @@ const tabletStyles = css`
 
   .k-CrowdfundingCard__progressBar__percent {
     display: inherit;
-  }
-
-  &.k-CrowdfundingCardWidget {
-    .k-CrowdfundingCard__title {
-      padding: 0 ${COMPONENT_GUTTER} 0 0;
-    }
   }
 
   &.k-CrowdfundingCard--titlesMinHeight .k-CrowdfundingCard__titleAndDesc {