+++ assets/javascripts/kitten/components/carousel/nav-tab-carousel/components/next.js
@@ -7,7 +7,11 @@ import {
 
 export const Next = ({ children, hoverColor }) => {
   return (
-    <StyledRightArrowContainer hoverColor={hoverColor}>
+    <StyledRightArrowContainer
+      hoverColor={hoverColor}
+      as="button"
+      className="k-u-reset-button"
+    >
       <StyledNextText>{children}</StyledNextText>
       <StyledArrowIcon fill="#fff" />
     </StyledRightArrowContainer>