+++ assets/javascripts/kitten/components/date-picker/components/navbar.js
@@ -51,14 +51,14 @@ export class Navbar extends Component {
           onClick={() => onPreviousClick()}
           left
         >
-          <ArrowIcon fill={iconColor} direction="left" />
+          <ArrowIcon color={iconColor} direction="left" />
         </StyledArrowIcon>
         <StyledArrowIcon
           aria-label={labels.nextMonth}
           onClick={() => onNextClick()}
           right
         >
-          <ArrowIcon fill={iconColor} />
+          <ArrowIcon color={iconColor} />
         </StyledArrowIcon>
       </div>
     )