+++ assets/javascripts/kitten/components/uploaders/document-manager/index.js
@@ -13,21 +13,41 @@ import { ClockCircleIcon } from '../../../components/icons/clock-circle-icon'
 import { Loader } from '../../../components/loaders/loader'
 import { DocumentIconEmpty } from '../../../components/icons/document-icon-empty'
 import { VisuallyHidden } from '../../../components/accessibility/visually-hidden'
+import { ScreenConfig } from '../../../constants/screen-config'
 
 const StyledDocumentUploader = styled.div`
-   {
-    input[type='file'] {
-      border: 0;
-      clip-path: inset(100%);
-      height: 1px;
-      overflow: hidden;
-      padding: 0;
-      position: absolute !important;
-      white-space: nowrap;
-      width: 1px;
-    }
+  input[type='file'] {
+    border: 0;
+    clip-path: inset(100%);
+    height: 1px;
+    overflow: hidden;
+    padding: 0;
+    position: absolute !important;
+    white-space: nowrap;
+    width: 1px;
+  }
 
-    input[type='file']:focus + label {
+  input[type='file']:focus + label {
+    outline: ${COLORS.primary4} solid ${pxToRem(2)};
+    outline-offset: ${pxToRem(2)};
+  }
+
+  input[type='file']:focus,
+  input[type='file']:hover {
+    & + label {
+      background-color: ${COLORS.background2};
+      border-color: ${COLORS.line1};
+      color: ${COLORS.font1};
+
+      svg,
+      path {
+        fill: ${COLORS.font1};
+      }
+    }
+  }
+
+  input[type='file']:active {
+    & + label {
       background-color: ${COLORS.primary2};
       border-color: ${COLORS.primary2};
       color: ${COLORS.background1};
@@ -37,41 +57,44 @@ const StyledDocumentUploader = styled.div`
         fill: ${COLORS.background1};
       }
     }
+  }
 
-    input[type='file']:disabled + label {
-      border-color: ${COLORS.line2};
-      background-color: ${COLORS.line2};
-      color: ${COLORS.background1};
-      pointer-events: none;
+  input[type='file']:disabled + label {
+    border-color: ${COLORS.line2};
+    background-color: ${COLORS.line2};
+    color: ${COLORS.background1};
+    pointer-events: none;
 
-      svg,
-      path {
-        fill: ${COLORS.background1};
-      }
+    svg,
+    path {
+      fill: ${COLORS.background1};
     }
+  }
 
-    .k-DocumentManager__uploader__button {
-      padding: ${pxToRem(20)};
-    }
+  .k-DocumentManager__uploader__button {
+    padding: ${pxToRem(20)};
+  }
 
-    .k-DocumentManager__uploader__container {
-      display: flex;
-      justify-content: stretch;
-      align-items: center;
-      width: 100%;
+  .k-DocumentManager__uploader__container {
+    display: flex;
+    justify-content: stretch;
+    align-items: center;
+    width: 100%;
+  }
+  .k-DocumentManager__uploader__documentIcon {
+    flex: 0 0 auto;
+    margin-right: ${pxToRem(20)};
+    align-self: flex-start;
+    @media (max-width: ${pxToRem(ScreenConfig.XS.max)}) {
+      display: none;
     }
-    .k-DocumentManager__uploader__documentIcon {
-      flex: 0 0 auto;
-      margin-right: ${pxToRem(20)};
-      align-self: flex-start;
-    }
-    .k-DocumentManager__uploader__content {
-      flex: 1 1 auto;
-    }
-    .k-DocumentManager__uploader__uploadIcon {
-      margin-left: ${pxToRem(20)};
-      flex: 0 0 auto;
-    }
+  }
+  .k-DocumentManager__uploader__content {
+    flex: 1 1 auto;
+  }
+  .k-DocumentManager__uploader__uploadIcon {
+    margin-left: ${pxToRem(20)};
+    flex: 0 0 auto;
   }
 `
 
@@ -107,6 +130,12 @@ const StyledDocumentDisplay = styled.div`
     &:focus {
       color: ${COLORS.primary3};
     }
+
+    &:focus {
+      outline: ${COLORS.primary4} solid ${pxToRem(2)};
+      outline-offset: ${pxToRem(2)};
+    }
+
     &:active {
       color: ${COLORS.primary1};
       transition: none;
@@ -260,7 +289,7 @@ export const DocumentManager = ({
         />
         <Button
           fluid
-          borderRadius={4}
+          borderRadius={6}
           {...buttonProps}
           as="label"
           htmlFor={id}
@@ -275,7 +304,6 @@ export const DocumentManager = ({
                 {documentIcon}
               </IconContainer>
             </div>
-
             <div className="k-DocumentManager__uploader__content">
               <Text
                 tag="p"
@@ -322,7 +350,7 @@ export const DocumentManager = ({
       >
         <Button
           fluid
-          borderRadius={4}
+          borderRadius={6}
           {...buttonProps}
           as="div"
           className={classNames(