+++ src/components/carousel/nav-tab-carousel/components/pagination.js
@@ -24,11 +24,11 @@ var _range = require("./../../../../helpers/utils/range");
 var BulletPointStyles = _styledComponents.default.div.withConfig({
   displayName: "pagination__BulletPointStyles",
   componentId: "sc-1thdyq2-0"
-})(["min-width:", ";min-height:", ";margin-left:", ";margin-right:", ";background-color:", ";", ""], (0, _typography.pxToRem)(6), (0, _typography.pxToRem)(6), (0, _typography.pxToRem)(4), (0, _typography.pxToRem)(4), _colorsConfig.default.background1, function (_ref) {
+})(["min-width:", ";min-height:", ";margin-left:", ";margin-right:", ";background-color:", ";", " &:focus{outline:", " solid ", ";outline-offset:", ";}"], (0, _typography.pxToRem)(6), (0, _typography.pxToRem)(6), (0, _typography.pxToRem)(4), (0, _typography.pxToRem)(4), _colorsConfig.default.background1, function (_ref) {
   var isSelected = _ref.isSelected,
       activeColor = _ref.activeColor;
   return isSelected && (0, _styledComponents.css)(["background-color:", ";"], activeColor);
-});
+}, _colorsConfig.default.primary3, (0, _typography.pxToRem)(2), (0, _typography.pxToRem)(2));
 
 var PaginationStyles = _styledComponents.default.div.withConfig({
   displayName: "pagination__PaginationStyles",
@@ -47,13 +47,17 @@ var Pagination = function Pagination(_ref2) {
       href: link,
       key: index,
       isSelected: activeIndex === index + 1,
-      activeColor: activeColor
+      "aria-current": activeIndex === index + 1,
+      activeColor: activeColor,
+      tabIndex: "0"
     });
   }), hasNoLinks && (0, _range.createRangeFromZeroTo)(totalIndex).map(function (_, index) {
     return /*#__PURE__*/_react.default.createElement(BulletPointStyles, {
       key: index,
       isSelected: activeIndex === index + 1,
-      activeColor: activeColor
+      "aria-current": activeIndex === index + 1,
+      activeColor: activeColor,
+      tabIndex: "0"
     });
   }));
 };