+++ src/components/form/text-input-with-button/index.js
@@ -58,7 +58,7 @@ var StyledTextInput = (0, _styledComponents.default)(_textInput.TextInput).withC
 var StyledButton = _styledComponents.default.button.withConfig({
   displayName: "text-input-with-button__StyledButton",
   componentId: "sc-1yqni7d-2"
-})(["", ";font-size:", ";line-height:1.3;flex:1 0 auto;outline:none;appearance:none;cursor:pointer;padding:0 ", ";border-radius:0;", ";", " ", " input:invalid:not(:focus) + &{background-color:", ";border-color:", ";}", ""], _typographyConfig.default.fontStyles.regular, (0, _typography.stepToRem)(-1), (0, _typography.pxToRem)(30), function (_ref) {
+})(["", ";font-size:", ";line-height:1.3;flex:1 0 auto;appearance:none;cursor:pointer;padding:0 ", ";border-radius:0;", ";", " ", " input:invalid:not(:focus) + &{background-color:", ";border-color:", ";}", ""], _typographyConfig.default.fontStyles.regular, (0, _typography.stepToRem)(-1), (0, _typography.pxToRem)(30), function (_ref) {
   var modifier = _ref.modifier;
   return (0, _modifierStyles.modifierStyles)(modifier);
 }, function (_ref2) {