+++ src/esm/components/form/text-input/index.js
@@ -21,7 +21,7 @@ import classNames from 'classnames';
 var StyledInput = styled.input.withConfig({
   displayName: "text-input__StyledInput",
   componentId: "sc-11wej6v-0"
-})(["--input-padding-horizontal:", ";--input-border-width:", ";font-size:", ";line-height:1em;", ";box-sizing:border-box;border-width:", ";border-width:var(--input-border-width);border-style:solid;border-radius:0;width:100%;appearance:none;background-color:", ";color:", ";border-color:", ";outline:none;::placeholder{color:", ";}height:", ";&.k-Form-TextInput--tiny{height:", ";}&.k-Form-TextInput--huge{height:", ";@media (min-width:", "px){height:", ";font-size:", ";}}&.k-Form-TextInput--giant{height:", ";@media (min-width:", "px){height:", ";font-size:", ";}}padding:", " ", ";padding:", " var(--input-padding-horizontal);&.k-Form-TextInput--orion{border-radius:", ";height:", ";@media (min-width:", "px){--input-padding-horizontal:", ";border-radius:", ";height:", ";font-size:", ";@media screen and (-ms-high-contrast:active),(-ms-high-contrast:none){padding:", " ", ";}}}&.k-Form-TextInput--alignCenter{text-align:center;}:disabled{color:", ";border-color:", ";background-color:", ";cursor:not-allowed;}&.k-Form-TextInput--valid{color:", ";border-color:", ";}&.k-Form-TextInput--error{color:", ";border-color:", ";}&:invalid{box-shadow:none;outline:none;color:", ";border-color:", ";}&:focus{outline:none;color:", ";border-color:", ";}&.k-Form-TextInput-hasDigits{width:calc( var(--input-content-width) * 1ch + (2 * (var(--input-padding-horizontal) + var(--input-border-width))) );}&.k-Form-TextInput-hasDigits_2{text-align:center;}@media screen and (-ms-high-contrast:active),(-ms-high-contrast:none){&.k-Form-TextInput-hasDigits_2{text-align:center;width:calc(2em + ", ");&.k-Form-TextInput--orion{width:calc(2em + ", ");@media (min-width:", "px){width:calc(2em + ", ");}}}&.k-Form-TextInput-hasDigits_4{width:calc(4em + ", ");&.k-Form-TextInput--orion{width:calc(4em + ", ");@media (min-width:", "px){width:calc(4em + ", ");}}}&.k-Form-TextInput-hasDigits_12{width:calc(12em + ", ");&.k-Form-TextInput--orion{width:calc(12em + ", ");@media (min-width:", "px){width:calc(12em + ", ");}}}}"], pxToRem(15), pxToRem(2), stepToRem(-1), TYPOGRAPHY.fontStyles.light, pxToRem(2), COLORS.background1, COLORS.font1, COLORS.line1, COLORS.font2, pxToRem(50), pxToRem(40), pxToRem(70), ScreenConfig.M.min, pxToRem(80), stepToRem(0), pxToRem(70), ScreenConfig.M.min, pxToRem(90), stepToRem(0), pxToRem(10), pxToRem(15), pxToRem(10), pxToRem(6), pxToRem(60), ScreenConfig.M.min, pxToRem(30), pxToRem(8), pxToRem(70), stepToRem(0), pxToRem(10), pxToRem(30), COLORS.font2, COLORS.line1, COLORS.line1, COLORS.tertiary2, COLORS.tertiary2, COLORS.error3, COLORS.error3, COLORS.error3, COLORS.error3, COLORS.font1, COLORS.line2, pxToRem(2 * (10 + 2)), pxToRem(2 * (15 + 2)), ScreenConfig.M.min, pxToRem(2 * (30 + 2)), pxToRem(2 * (10 + 2)), pxToRem(2 * (15 + 2)), ScreenConfig.M.min, pxToRem(2 * (30 + 2)), pxToRem(2 * (10 + 2)), pxToRem(2 * (15 + 2)), ScreenConfig.M.min, pxToRem(2 * (30 + 2)));
+})(["--input-padding-horizontal:", ";--input-border-width:", ";font-size:", ";line-height:1em;", ";box-sizing:border-box;border-width:", ";border-width:var(--input-border-width);border-style:solid;border-radius:0;width:100%;appearance:none;background-color:", ";color:", ";border-color:", ";::placeholder{color:", ";}&.k-Form-TextInput--tiny{height:", ";}&.k-Form-TextInput--regular{height:", ";}&.k-Form-TextInput--huge{height:", ";@media (min-width:", "px){height:", ";font-size:", ";}}&.k-Form-TextInput--giant{height:", ";@media (min-width:", "px){height:", ";font-size:", ";}}padding:", " ", ";padding:", " var(--input-padding-horizontal);&.k-Form-TextInput--orion{&.k-Form-TextInput--regular{border-radius:", ";height:", ";}&.k-Form-TextInput--big{border-radius:", ";height:", ";@media (min-width:", "px){height:", ";--input-padding-horizontal:", ";border-radius:", ";font-size:", ";@media screen and (-ms-high-contrast:active),(-ms-high-contrast:none){padding:", " ", ";}}}}&.k-Form-TextInput--alignCenter{text-align:center;}:disabled{color:", ";border-color:", ";background-color:", ";cursor:not-allowed;}&.k-Form-TextInput--valid{color:", ";border-color:", ";}&.k-Form-TextInput--error{color:", ";border-color:", ";}&:invalid{box-shadow:none;color:", ";border-color:", ";}&:focus{color:", ";border-color:", ";outline:", " solid ", ";outline-offset:", ";}&.k-Form-TextInput-hasDigits{width:calc( var(--input-content-width) * 1ch + (2 * (var(--input-padding-horizontal) + var(--input-border-width))) );}&.k-Form-TextInput-hasDigits_2{text-align:center;}@media screen and (-ms-high-contrast:active),(-ms-high-contrast:none){&.k-Form-TextInput-hasDigits_2{text-align:center;width:calc(2em + ", ");&.k-Form-TextInput--orion{width:calc(2em + ", ");@media (min-width:", "px){width:calc(2em + ", ");}}}&.k-Form-TextInput-hasDigits_4{width:calc(4em + ", ");&.k-Form-TextInput--orion{width:calc(4em + ", ");@media (min-width:", "px){width:calc(4em + ", ");}}}&.k-Form-TextInput-hasDigits_12{width:calc(12em + ", ");&.k-Form-TextInput--orion{width:calc(12em + ", ");@media (min-width:", "px){width:calc(12em + ", ");}}}}"], pxToRem(15), pxToRem(2), stepToRem(-1), TYPOGRAPHY.fontStyles.light, pxToRem(2), COLORS.background1, COLORS.font1, COLORS.line1, COLORS.font2, pxToRem(40), pxToRem(50), pxToRem(70), ScreenConfig.M.min, pxToRem(80), stepToRem(0), pxToRem(70), ScreenConfig.M.min, pxToRem(90), stepToRem(0), pxToRem(10), pxToRem(15), pxToRem(10), pxToRem(4), pxToRem(50), pxToRem(6), pxToRem(60), ScreenConfig.M.min, pxToRem(70), pxToRem(30), pxToRem(8), stepToRem(0), pxToRem(10), pxToRem(30), COLORS.font2, COLORS.line1, COLORS.line1, COLORS.tertiary2, COLORS.tertiary2, COLORS.error3, COLORS.error3, COLORS.error3, COLORS.error3, COLORS.font1, COLORS.line2, COLORS.primary4, pxToRem(2), pxToRem(2), pxToRem(2 * (10 + 2)), pxToRem(2 * (15 + 2)), ScreenConfig.M.min, pxToRem(2 * (30 + 2)), pxToRem(2 * (10 + 2)), pxToRem(2 * (15 + 2)), ScreenConfig.M.min, pxToRem(2 * (30 + 2)), pxToRem(2 * (10 + 2)), pxToRem(2 * (15 + 2)), ScreenConfig.M.min, pxToRem(2 * (30 + 2)));
 var StyledTextareaContainer = styled.div.withConfig({
   displayName: "text-input__StyledTextareaContainer",
   componentId: "sc-11wej6v-1"
@@ -49,6 +49,8 @@ export var TextInput = /*#__PURE__*/function (_PureComponent) {
           name = _this$props.name,
           digits = _this$props.digits,
           tiny = _this$props.tiny,
+          regular = _this$props.regular,
+          big = _this$props.big,
           huge = _this$props.huge,
           giant = _this$props.giant,
           center = _this$props.center,
@@ -56,7 +58,7 @@ export var TextInput = /*#__PURE__*/function (_PureComponent) {
           variant = _this$props.variant,
           className = _this$props.className,
           style = _this$props.style,
-          others = _objectWithoutProperties(_this$props, ["valid", "error", "disabled", "name", "digits", "tiny", "huge", "giant", "center", "tag", "variant", "className", "style"]);
+          others = _objectWithoutProperties(_this$props, ["valid", "error", "disabled", "name", "digits", "tiny", "regular", "big", "huge", "giant", "center", "tag", "variant", "className", "style"]);
 
       var digitsClass = !!digits ? "k-Form-TextInput-hasDigits k-Form-TextInput-hasDigits_".concat(digits) : null;
 
@@ -74,9 +76,6 @@ export var TextInput = /*#__PURE__*/function (_PureComponent) {
             'k-Form-TextInput--valid': valid,
             'k-Form-TextInput--error': error,
             'k-Form-TextInput--disabled': disabled,
-            'k-Form-TextInput--tiny': tiny,
-            'k-Form-TextInput--huge': huge,
-            'k-Form-TextInput--giant': giant,
             'k-Form-TextInput--alignCenter': center
           }),
           style: !!digits ? _extends({
@@ -97,6 +96,8 @@ export var TextInput = /*#__PURE__*/function (_PureComponent) {
             'k-Form-TextInput--error': error,
             'k-Form-TextInput--disabled': disabled,
             'k-Form-TextInput--tiny': tiny,
+            'k-Form-TextInput--regular': regular,
+            'k-Form-TextInput--big': big,
             'k-Form-TextInput--huge': huge,
             'k-Form-TextInput--giant': giant,
             'k-Form-TextInput--alignCenter': center
@@ -116,6 +117,8 @@ TextInput.propTypes = {
   valid: PropTypes.bool,
   error: PropTypes.bool,
   tiny: PropTypes.bool,
+  regular: PropTypes.bool,
+  big: PropTypes.bool,
   huge: PropTypes.bool,
   giant: PropTypes.bool,
   center: PropTypes.bool,
@@ -130,6 +133,8 @@ TextInput.defaultProps = {
   valid: false,
   error: false,
   tiny: false,
+  regular: true,
+  big: false,
   huge: false,
   giant: false,
   center: false,