+++ src/esm/components/navigation/nav-bar/index.js
@@ -12,7 +12,7 @@ import { ScreenConfig } from '../../../constants/screen-config';
 var StyledNavBar = styled.div.withConfig({
   displayName: "nav-bar__StyledNavBar",
   componentId: "sc-1bcathh-0"
-})(["width:auto;flex-grow:1;.k-NavBar__nav{width:auto;margin:0;padding:0 ", ";background:", ";display:flex;justify-content:center;@media (min-width:", "){padding:0 ", ";}&::before,&::after{content:\"\";flex-grow:1;}}.k-NavBar__list{list-style-type:none;display:flex;align-items:center;margin:0;padding:0;}.k-NavBar__listItem + .k-NavBar__listItem{margin-left:", ";@media (min-width:", "){margin-left:", ";}}.k-NavBar__link{height:", ";box-sizing:border-box;", " font-size:", ";text-decoration:none;text-align:center;display:flex;align-items:center;justify-content:center;flex-direction:row;position:relative;outline:none;color:", ";border-bottom:", " solid ", ";transition:color .2s,border-color .2s;cursor:pointer;&:hover{color:", ";text-decoration:none;}&:focus,&[aria-current=\"page\"]{border-color:", ";color:", ";border-width:", ";}@media (min-width:", "){height:", ";}}&.k-NavBar--big .k-NavBar__link{height:", ";@media (min-width:", "){height:", ";}}"], pxToRem(CONTAINER_PADDING_THIN), function (_ref) {
+})(["width:auto;flex-grow:1;.k-NavBar__nav{width:auto;margin:0;padding:0 ", ";background:", ";display:flex;justify-content:center;@media (min-width:", "){padding:0 ", ";}&::before,&::after{content:\"\";flex-grow:1;}}.k-NavBar__list{list-style-type:none;display:flex;align-items:center;margin:0;padding:0;}.k-NavBar__listItem + .k-NavBar__listItem{margin-left:", ";@media (min-width:", "){margin-left:", ";}}.k-NavBar__link{height:", ";box-sizing:border-box;", " font-size:", ";text-decoration:none;text-align:center;display:flex;align-items:center;justify-content:center;flex-direction:row;position:relative;color:", ";border-bottom:", " solid ", ";transition:color .2s,border-color .2s;cursor:pointer;&:hover{color:", ";text-decoration:none;}&:focus{outline:", " solid ", ";outline-offset:", ";}&:focus,&[aria-current=\"page\"]{border-color:", ";color:", ";border-width:", ";}@media (min-width:", "){height:", ";}}&.k-NavBar--big .k-NavBar__link{height:", ";@media (min-width:", "){height:", ";}}"], pxToRem(CONTAINER_PADDING_THIN), function (_ref) {
   var colors = _ref.colors;
   return colors.background || COLORS.primary6;
 }, pxToRem(ScreenConfig.S.min), pxToRem(CONTAINER_PADDING), pxToRem(25), pxToRem(ScreenConfig.S.min), pxToRem(50), pxToRem(65), TYPOGRAPHY.fontStyles.regular, stepToRem(-1), function (_ref2) {
@@ -24,7 +24,7 @@ var StyledNavBar = styled.div.withConfig({
 }, function (_ref4) {
   var colors = _ref4.colors;
   return colors.activeLink || COLORS.primary1;
-}, function (_ref5) {
+}, COLORS.primary4, pxToRem(2), pxToRem(2), function (_ref5) {
   var colors = _ref5.colors;
   return colors.activeBorder || COLORS.primary1;
 }, function (_ref6) {