+++ styleguide/component/presentation-elements/hero.js
@@ -1,6 +1,0 @@
-import React from 'react'
-import { KarlHeroExamples } from '../../../assets/javascripts/kitten/karl/heroes/hero'
-
-export default () => {
-  return <KarlHeroExamples />
-}